refactor: extract Dashboard subhead component
